+++ src/components/base/b-list/b-list.ts
@@ -322,6 +322,15 @@ class bList extends bListProps implements iVisible, iWidth, iActiveItems {
 		return iItems.getItemKey(this, item, i);
 	}
 
+	/**
+	 * Returns href for item. Is used at the rendering stage
+	 *
+	 * @param item
+	 */
+	protected getHref(item: this['Item']): CanUndef<string> {
+		return item.href;
+	}
+
 	protected override initModEvents(): void {
 		super.initModEvents();
 		iVisible.initModEvents(this);